requirements.txt
celery==4.4.0
requests==2.32.2
watchdog==2.1.1
argh==0.26.2
websockets==8.1
aiofiles==0.4.0
pyyaml==5.3.1
tqdm==4.66.5
onnxruntime-gpu==1.18.0
transformers==4.29.2
diffusers==0.21.4
numba==0.60.0
numpy==1.26.4
opencv_contrib_python==4.10.0.84
opencv_python_headless==4.10.0.84
opencv_python==4.10.0.84
bchlib==2.1.3
stn==1.0.1
Pillow==10.4.0
reedsolo==1.7.0
scikit-learn==1.2.2
timm==0.9.5
datasets==3.0.0
compressai==1.2.6
scikit-image==0.24.0
pycryptodome==3.20.0
accelerate==0.34.2
galois==0.4.2
ldpc==0.1.53
scipy==1.13.1
open_clip_torch==2.23.0
clip @ git+https://github.com/openai/CLIP.git@a1d071733d7111c9c014f024669f959182114e33